feat: add split mfe-api-kit to "core" and "react-core"-packages
