+++ Readme.md
@@ -27,7 +27,7 @@ lerna run start --scope="@mf/*"
 - `@mfe/users` - отвечает за что-то типо юзеров
 
 ## Утилиты
-- `@app-lib/mfe-api-kit` - библиотека с примитивами для контрактов
+- `@app-lib/react-core` - библиотека с примитивами для контрактов
 - `@app-lib/ui-kit` - ну это типо ui-kit 
 
 ## Чепуха
