+++ app/host/package.json
@@ -6,7 +6,7 @@
     "piral"
   ],
   "dependencies": {
-    "@app-lib/mfe-api-kit": "*",
+    "@app-lib/react-core": "*",
     "@app-lib/ui-kit": "*",
     "@evo.contracts/tenants": "1.0.0",
     "@tanstack/react-query": "^5.77.2",
@@ -32,7 +32,7 @@
       "react": ".",
       "react-dom": ".",
       "@tanstack/react-query": ".",
-      "@app-lib/mfe-api-kit": ".",
+      "@app-lib/react-core": ".",
       "@app-lib/ui-kit": "."
     },
     "inherit": [
