+++ contracts/license/package.json
@@ -4,6 +4,7 @@
   "module": "./dist/index.mjs",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
+  "type": "module",
   "scripts": {
     "build": "vite build"
   },
@@ -12,22 +13,19 @@
   "license": "ISC",
   "description": "",
   "peerDependencies": {
-    "@app-lib/mfe-api-kit": "*",
+    "@app-lib/react-core": "*",
     "react": "^18",
     "react-dom": "^18"
   },
   "devDependencies": {
-    "@app-lib/mfe-api-kit": "*",
+    "@app-lib/react-core": "*",
     "@app-system/config": "*",
-    "@rollup/plugin-replace": "^6.0.2",
     "@types/node": "^22",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "@vitejs/plugin-react": "^4.4.1",
     "react": "^18",
     "react-dom": "^18",
     "typescript": "^5.8.3",
-    "vite": "^6.3.5",
-    "vite-plugin-dts": "^4.5.3"
+    "vite": "^6.3.5"
   }
 }
