+++ lib/core/src/domain/common/eventNamingStrategy.ts
