+++ lib/core/src/domain/request-api/index.ts
@@ -1,0 +1,3 @@
+export * from './exposeApi';
+export type { ProcedureResolverOptions } from './apiRequestBuilder';
+export * from './clientTypes';
