+++ lib/core/src/utils/index.ts
@@ -1,0 +1,3 @@
+export function wait(ms: number): Promise<void> {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
