+++ lib/core/tsconfig.json
