--- lib/mfe-api-kit/src/request-api/index.ts
@@ -1,3 +1,0 @@
-export * from './exposeApi';
-export * from './createReactQueryClient';
-export type { ProcedureResolverOptions } from './core/apiRequestBuilder';
