--- lib/mfe-api-kit/src/utils.ts
@@ -1,3 +1,0 @@
-export function wait(ms: number): Promise<void> {
-    return new Promise(resolve => setTimeout(resolve, ms));
-}
