+++ lib/react-core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@app-lib/mfe-api-kit",
+  "name": "@app-lib/react-core",
   "version": "1.0.0",
   "type": "module",
   "scripts": {
@@ -10,6 +10,7 @@
   "types": "./dist/index.d.ts",
   "peerDependencies": {
     "@apollo/client": "^3.13.8",
+    "@app-lib/core": "1.0.0",
     "@tanstack/react-query": "^5.76.1",
     "react": "^18",
     "react-dom": "^18"
@@ -17,17 +18,14 @@
   "devDependencies": {
     "@apollo/client": "^3.13.8",
     "@app-system/config": "*",
-    "@rollup/plugin-replace": "^6.0.2",
     "@tanstack/react-query": "^5.76.1",
     "@types/node": "^22",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "@vitejs/plugin-react": "^4.4.1",
     "react": "^18",
     "react-dom": "^18",
     "typescript": "^5.8.3",
-    "vite": "^6.3.5",
-    "vite-plugin-dts": "^4.5.3"
+    "vite": "^6.3.5"
   },
   "keywords": [],
   "author": "",
