+++ lib/react-core/src/index.ts
