+++ lib/react-core/src/request-api/index.ts
@@ -1,0 +1,3 @@
+export type { ProcedureResolverOptions, ApiRegister } from '@app-lib/core';
+export { exposeApi } from '@app-lib/core';
+export * from './createReactQueryClient';
