+++ lib/react-core/vite.config.ts
