+++ lib/ui-kit/package.json
@@ -4,6 +4,7 @@
   "module": "./dist/index.mjs",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
+  "type": "module",
   "scripts": {
     "build": "vite build",
     "storybook": "storybook dev -p 6006",
